fix: fixed display precision

+++ software/src/Interfaces/Frontend/BMS.hpp
@@ -8,8 +8,8 @@ namespace fake {
 
 class BMS : public QObject, public FakeInterface {
     Q_OBJECT
-    Q_PROPERTY(QList<int> voltages MEMBER m_voltages NOTIFY voltagesChanged)
-    Q_PROPERTY(QList<int> resistances MEMBER m_resistances NOTIFY resistancesChanged)
+    Q_PROPERTY(QList<float> voltages MEMBER m_voltages NOTIFY voltagesChanged)
+    Q_PROPERTY(QList<float> resistances MEMBER m_resistances NOTIFY resistancesChanged)
   public:
     BMS(const std::string& /* dbc_file_path */ = "") : QObject(nullptr), m_voltages(5 * 28, -40), m_resistances(5 * 28, -40) {
         this->FakeInterface::startReceiving();
@@ -60,8 +60,8 @@ class BMS : public QObject, public FakeInterface {
     void resistancesChanged();
 
   private:
-    QList<int> m_voltages;
-    QList<int> m_resistances;
+    QList<float> m_voltages;
+    QList<float> m_resistances;
 
   private:
     void generateValues();
